UART Transmitter
The UART Transmitter block takes inputs clk, resetn, uart_tx_en, uart_tx_data and gives output uart_txd and uart_tx_busy.  
Basys-3 board runs at clock frequency of 100MHz so we declared it as a parameter and we are using Bluetooth module so the baudrate for transmission is 9600.  
Now we are required to calculate the number of clock cycles required to transmit a single bit so it can be calculated by dividing clock frequency with baudrate.  

 
Now we need to count upto this number so the number of bits required to calculate the CYCLES_PER_BIT can be found by using the formula “ 1+$clog2(CYCLES_PER_BIT) “ .  
We have declared some wires like next_bit, pay_load_done, stop_done.  
next_bit will become 1 when the count reaches to number of cycles required to transmit a single bit. 
payload_done will become 1 when it reaches the total number of payload that is going to be transmitted.  
stop_done will become 1 when the present state reaches STOP state. 
There are 4 states in this block i.e IDLE, START, SEND, STOP. For every positive edge clock whenever there is no reset the present state changes to the next state else it stays in the IDLE state. 
In Next State logic when the p_state is IDLE it will reaches to next state START whenever uart_tx_en will become high else it remains in the same state. When the p_state is START it reaches next state SEND whenever next_bit is high else it remains in the same state. When the p_state is SEND it reaches next state STOP when the payload_done is high else it remains in the same state. When the p_state in STOP it reaches the initial state IDLE whenever stop_done signal is high else it remains in the same state. 
Whenever there is reset then the data_to_send register is filled with all 0s.  
Whenever p_state is IDLE and the uart_tx_en is high then the data_to_send register is filled with the input data.  
Whenever p_state in SEND state and next_bit is high then the data in the data_to_send register will right shift to 1 bit. 
bit_counter increments its value whenever it is in SEND state and the next_bit is in high otherwise the bit_counter will reset to 0. 
cycle_counter increments its value whenever p_state is in START, SEND and STOP states. 
txd_reg is initially in high state whenever it starts sending the data the first bit it sends as 0 in the significance of data being transmitting. txd_reg is filled with the least significant bit(lsb) of data_to_send register whenever p_state is in SEND state.