UART Receiver
The UART Transmitter block takes inputs clk, resetn, uart_rxd, uart_rx_en, uart_rx_break and gives output uart_rx_valid and uart_rx_data.  
Basys-3 board runs at clock frequency of 100MHz so we declared it as a parameter and we are using Bluetooth module so the baudrate for transmission is 9600.  
Now we are required to calculate the number of clock cycles required to transmit a single bit so it can be calculated by dividing clock frequency with baudrate.  
Now we need to count upto this number so the number of bits required to calculate the CYCLES_PER_BIT can be found by using the formula “ 1+$clog2(CYCLES_PER_BIT) “ .  
We have declared some wires like next_bit, pay_load_done.  
next_bit will become 1 when the count reaches to number of cycles required to transmit a single bit. 
payload_done will become 1 when it reaches the total number of payload that is going to be transmitted.  
There are 4 states in this block i.e IDLE, START, RECV, STOP. For every positive edge clock whenever there is no reset the present state changes to the next state else it stays in the IDLE state. 
In Next State logic when the p_state is IDLE it will reaches to next state START whenever rxd_reg will become high else it remains in the same state. When the p_state is START it reaches next state RECV whenever next_bit is high else it remains in the same state. When the p_state is RECV it reaches next state STOP when the payload_done is high else it remains in the same state. When the p_state in STOP it reaches the initial state IDLE whenever next_bit is high else it remains in the same state. 
bit_counter increments its value whenever it is in RECV state and the next_bit is in high otherwise the bit_counter will reset to 0. 
cycle_counter increments its value whenever p_state is in START, RECV and STOP states. 
uart_rx_break signal will high when the input data is all 0s. 
uart_rx_valid signal will high when the p_state is STOP and n_state is IDLE. 
bit_sample is loaded with the rxd_reg whenever cycle_counter reaches half of the  cycles_per_bit value. 
rxd_reg is loaded with uart_rxd value whenever clock edge comes and resetn is high and uart_rx_en is high. 
received_data register msb bit is loaded with the bit sample value whenever p_state is RECV and next_bit is high. Each time this received_data register is right shifts its value by 1 bit