Top Module
In this UART_top top module inputs are clk, uart_rxd, resetn and the outputs are uart_txd, data and LA(Left Anode), LC(Left Cathode), RA(Right Anode), RC(Right Cathode) which are used to control the motors.  
We have used some wires like uart_rx_data, uart_rx_valid, uart_rx_break, uart_tx_busy, uart_tx_data, uart_tx_en and one data_reg register to store the received data. These wires and registers are corresponding to the UART_Rx and UART_Tx modules.  
Whenever there is no reset and uart_rx_valid is high then data_reg is loaded with the uart_rx_data. The output data is loaded with the data_reg value. 

 
Depending upon the received data in the data signal we will control the motors. 
If the received data is 0 then the motors will stop working. 
If the received data is 1 then the motors move in forward direction. 
If the received data is 2 then the motors move in backward direction. 
If the received data is 3 then the motors turns right direction. 
If the received data is 4 then the motors turns left direction.  
If the received data is 5 then the motors move in circular direction. 
 
Note:- The data which is going to be received in the FPGA is sent through by a mobile application in which all the prerequisites are preinstalled in that mobile application. 
The mobile application is implemented using MIT App Inventor website.